Fixed scientific notation bug.
Many OBS applications save cells with long numbers in a loose scientific notation format, this means any cell with a float or double length number will be transformed into a relatively close but not correct scientific notation. When reading this file later if you attempt to use that number from the cell it will be completely different than what you actually put in it since it was saved with an incorrect scientific notation. The real number is saved as an office:value attribute attached to the table cell. If a office:value attribute is found on a cell we should use that value instead to avoid format bugs like long numbers.